+++ mount/index.ts
@@ -1,17 +1,16 @@
-import {ComponentType, createElement, render} from "preact";
+import {ComponentFactory, createElement, render} from "preact";
 import {mojave} from "../@types/mojave";
 import {find} from "../dom/traverse";
 import {extend} from "../extend";
 import {parseElementAsJson} from "../json";
 
 
-
-export function mount (selector: string, mountable: mojave.MountableFunction, options?: mojave.FunctionMountOptions): void;
-export function mount (selector: string, mountable: ComponentType<any>, options?: mojave.ComponentMountOptions): void;
-export function mount (selector: string, mountable: mojave.MountableClass, options?: mojave.ClassMountOptions): void;
-export function mount <T extends mojave.Mountable>(selector: string, mountable: mojave.Mountable, options?: mojave.MountOptions) : void
+/**
+ * Mounts a Preact function or class component into all elements matching the given selector.
+ */
+export function mountJsx<TPreactComponent extends ComponentFactory<any>>(selector: string, mountable: TPreactComponent, options?: mojave.ComponentMountOptions<TPreactComponent>): void
 {
-    doMount(find(selector), mountable, options);
+    find(selector).forEach(node => doMountJsx<TPreactComponent>(node, mountable, options));
 }
 
 
@@ -21,9 +20,9 @@ export function mount <T extends mojave.Mountable>(selector: string, mountable: 
  * The importer must import the component.
  * Example:
  *
- *     mountLazy(".selector", () => import("./src/MyComp"));
+ *     mountLazyJsx<MyPreactComp>(".selector", () => import("./src/MyPreactComp"));
  */
-export function mountLazy <T extends mojave.Mountable>(selector: string, importer: () => Promise<any>, options?: mojave.MountOptions) : void
+export function mountLazyJsx<TPreactComponent extends ComponentFactory<any>>(selector: string, importer: () => Promise<any>, options?: mojave.ComponentMountOptions<TPreactComponent>) : void
 {
     let elements = find(selector);
 
@@ -33,69 +32,143 @@ export function mountLazy <T extends mojave.Mountable>(selector: string, importe
     }
 
     importer().then(
-        module => doMount(elements, module.default, options),
+        module => {
+            elements.forEach(element => {
+                doMountJsx<TPreactComponent>(element, module.default, options);
+            });
+        },
         error => console.error(`Mounting of component of path '${selector}' failed: ${error.message}`, error)
     );
 }
 
 
-
 /**
- * Actually mounts on the given elements
+ * Mounts a Preact function or class component
  */
-function doMount (elements: HTMLElement[], mountable: mojave.Mountable, rawOptions?: mojave.MountOptions) : void
+function doMountJsx<TPreactComponent extends ComponentFactory<any>>(node: HTMLElement, mountable: TPreactComponent, options?: mojave.ComponentMountOptions<TPreactComponent>): void
 {
-    let mountableAny = mountable as any;
-    let options = extend({
-        type: "func",
-    }, rawOptions || {}) as mojave.MountOptions & {type: mojave.MountableType};
+    options = options || {};
+    let parent = node.parentElement;
+    let params = (options.params || {});
 
-    elements.forEach(
-        node =>
-        {
-            // check whether is a JSX component (i.e. it has no `init()` method).
-            if ("jsx" === options.type)
-            {
-                let opts = options as mojave.ComponentMountOptions;
-                let parent = node.parentElement;
-                let params = opts.params || {};
+    if (!parent)
+    {
+        console.error("Can't mount on container without parent.");
+        return;
+    }
 
-                if (!parent)
-                {
-                    console.error("Can't mount on container without parent.");
-                    return;
-                }
+    if (!options.hydrate)
+    {
+        // if the node should not be hydrated, try to use the content as JSON
+        params = extend(params, parseElementAsJson(node) || {});
 
-                if (!opts.hydrate)
-                {
-                    // if the node should not be hydrated, try to use the content as JSON
-                    params = extend(params, parseElementAsJson(node) || {});
+        // remove node before mount, so that we can ensure that preact doesn't reuse it.
+        parent.removeChild(node);
+    }
 
-                    // remove node before mount, so that we can ensure that preact doesn't reuse it.
-                    parent.removeChild(node);
-                }
-
-                // render
-                render(
-                    createElement(mountable as ComponentType<any>, params),
-                    parent,
-                    opts.hydrate ? node : undefined
-                );
-            }
-            else if ("class" === options.type)
-            {
-                let standaloneOptions = options as mojave.ClassMountOptions;
-                const mounted = new mountableAny(node, ...(standaloneOptions.params || []));
-                mounted.init();
-            }
-            else
-            {
-                let standaloneOptions = options as mojave.FunctionMountOptions;
-                (mountable as Function)(node, ...(standaloneOptions.params || []));
-            }
-        }
+    // render
+    render(
+        createElement(mountable, params),
+        parent,
+        options.hydrate ? node : undefined
     );
 }
 
 
+/**
+ * Mounts a StandaloneComponent into all elements matching the given selector.
+ */
+export function mountClass<TStandaloneComponent extends mojave.MountableClass>(selector: string, mountable: TStandaloneComponent, options?: mojave.ClassMountOptions<TStandaloneComponent>): void
+{
+    find(selector).forEach(node => {
+        doMountClass<TStandaloneComponent>(node, mountable, options);
+    });
+}
 
+
+/**
+ * Mounts the component lazily, if an element matching the selector exists.
+ *
+ * The importer must import the component.
+ * Example:
+ *
+ *     mountLazyClass<MyStandaloneComp>(".selector", () => import("./src/MyStandaloneComp"));
+ */
+export function mountLazyClass <TStandaloneComponent extends mojave.MountableClass>(selector: string, importer: () => Promise<any>, options?: mojave.ClassMountOptions<TStandaloneComponent>) : void
+{
+    let elements = find(selector);
+
+    if (!elements.length)
+    {
+        return;
+    }
+
+    importer().then(
+        module => {
+            elements.forEach(element => {
+                doMountClass<TStandaloneComponent>(element, module.default, options);
+            });
+        },
+        error => console.error(`Mounting of component of path '${selector}' failed: ${error.message}`, error)
+    );
+}
+
+
+/**
+ * Mounts a StandaloneComponent
+ */
+function doMountClass<TStandaloneComponent extends mojave.MountableClass>(node: HTMLElement, mountable: TStandaloneComponent, options?: mojave.ClassMountOptions<TStandaloneComponent>): void
+{
+    options = options || {};
+    const mounted = new mountable(node, ...(options.params || []));
+    mounted.init();
+}
+
+
+/**
+ * Mounts a function into all elements matching the given selector.
+ */
+export function mount<TFunction extends mojave.MountableFunction>(selector: string, mountable: TFunction, options?: mojave.FunctionMountOptions<TFunction>): void
+{
+    find(selector).forEach(node => {
+        doMountFunction<TFunction>(node, mountable, options);
+    });
+}
+
+
+/**
+ * Mounts the component lazily, if an element matching the selector exists.
+ *
+ * The importer must import the component.
+ * Example:
+ *
+ *     mountLazy<MyFunctionComp>(".selector", () => import("./src/MyFunctionComp"));
+ */
+export function mountLazy <TFunction extends mojave.MountableFunction>(selector: string, importer: () => Promise<any>, options?: mojave.FunctionMountOptions<TFunction>) : void
+{
+    let elements = find(selector);
+
+    if (!elements.length)
+    {
+        return;
+    }
+
+    importer().then(
+        module => {
+            elements.forEach(element => {
+                doMountFunction<TFunction>(element, module.default, options);
+            });
+        },
+        error => console.error(`Mounting of component of path '${selector}' failed: ${error.message}`, error)
+    );
+}
+
+
+/**
+ * Mounts a function
+ */
+function doMountFunction<TFunction extends mojave.MountableFunction>(node: HTMLElement, mountable: TFunction, options?: mojave.FunctionMountOptions<TFunction>): void
+{
+    options = options || {};
+    mountable(node, ...(options.params || []));
+}